refactor(cbh): use direct dependency paths and relocate StorageKey to cbh_model
Two post-fragmentation cleanups of the cbh_* crate family.

Honest dependency provenance (cbh_analysis, cbh_render): drop the crate-root
`use cbh_model as model;` / `use cbh_analysis as analyze;` aliases that let the
moved modules keep their old `crate::model::` / `crate::analyze::` import paths.
Those paths read as if `model`/`analyze` were local submodules when they are
external crates. Repoint every reference to the direct `cbh_model::` /
`cbh_analysis::` path, matching the convention the shell-derived crates already
use. `crate::analyze::` inside cbh_analysis is left alone -- there it is a
genuine internal module.

Relocate StorageKey + parse_key to cbh_model (from cbh_analysis): a
storage-key type and its parser lived in the analysis crate, away from
DiscriminantSet (which StorageKey contains) and the STORAGE_VERSION /
OBJECTS_SEGMENT constants parse_key validates against -- all in cbh_model.
parse_key is the exact inverse of DiscriminantSet's clean_key/dirty_key/
bless_key construction methods, so it now sits beside them in comparability.rs.
Consumers (cbh_analysis, cbh_analyze) import from cbh_model directly; the
cbh_analysis root no longer re-exports them.

Move the parse_key benchmark to cbh_model's storage_key group accordingly (and
drop cbh_analysis's now-empty bench and its criterion dev-dep). Fix a latent bug
in that benchmark: its input key omitted the `objects` segment, so parse_key
rejected it early and never measured the successful decomposition path.
