+++ packages/cbh_analysis/Cargo.toml
@@ -38,13 +38,8 @@ smallvec = { workspace = true, features = ["serde"] }
 thread_aware = { workspace = true, optional = true, features = ["derive"] }
 
 [dev-dependencies]
-criterion = { workspace = true }
 futures = { workspace = true, features = ["executor"] }
 mutants = { workspace = true }
-
-[[bench]]
-name = "cbh_analysis"
-harness = false
 
 [lints]
 workspace = true
