+++ packages/cbh_analysis/src/analyze/findings.rs
@@ -50,12 +50,12 @@ use std::ops::Range;
 use std::sync::Arc;
 
 use anyspawn::Spawner;
+use cbh_model::{BenchmarkId, DiscriminantSet, MetricKind};
 use cbh_stats as stats;
 use serde::Serialize;
 
 use crate::analyze::parallel::{balanced_chunk_sizes, worker_count};
 use crate::analyze::{Series, SeriesPoint};
-use crate::model::{BenchmarkId, DiscriminantSet, MetricKind};
 
 /// Tunable parameters of the engine-aware analysis.
 #[derive(Clone, Copy, Debug)]
@@ -1212,12 +1212,12 @@ mod tests {
 
     use std::sync::Arc;
 
+    use cbh_model::{DiscriminantSet, MetricKind};
     use jiff::Timestamp;
     use nonempty::nonempty;
 
     use super::*;
     use crate::analyze::{Blessing, SeriesPoint};
-    use crate::model::{DiscriminantSet, MetricKind};
 
     /// Builds a Callgrind-style series carrying `values` in topological order, with
     /// no dispersion (no confidence interval).
