+++ packages/cbh_analysis/src/analyze/signal_validation.rs
@@ -57,11 +57,11 @@
 
 use std::sync::Arc;
 
+use cbh_model::{BenchmarkId, DiscriminantSet, MetricKind};
 use nonempty::nonempty;
 
 use crate::analyze::findings::find_changes;
 use crate::analyze::{AnalysisConfig, AnalysisContext, AnalysisMode, Series, SeriesPoint};
-use crate::model::{BenchmarkId, DiscriminantSet, MetricKind};
 
 /// How a rise in the measured metric is judged.
 ///
